[CEN-1420] Avoid tests duplication on env basis
