+++ .gitignore
@@ -1,5 +1,6 @@
 certs
 k6
+.env*
 .env.development.local
 .env.test.local
 .env.production.local
