--- test/smoke/prod/env.production.local.sample
@@ -1,9 +1,0 @@
-APIM_SK=[your subscription key]
-APIM_RTDPRODUCT_SK=[subscription key under product RTD_API_Product]
-MAUTH_CERT_NAME=[name of the mauth certificate file in certs directory]
-MAUTH_PRIVATE_KEY_NAME=[name of the mauth rivate key file in certs directory ]
-PAN_OK=[pan of a card enrolled]
-FISCAL_CODE_EXISTING=[fiscal code]
-
-# Max content length for /hashed-pans API in bytes (0 means no max length)
-RTD_HASHPAN_MAX_CONTENT_LENGTH=0
