+++ test/smoke/rtdPaymentInstrumentManager.js
@@ -1,0 +1,60 @@
+import { group } from 'k6'
+import {
+	getHashedPan,
+	getSalt,
+} from '../common/api/rtdPaymentInstrumentManager.js'
+import { assert, statusOk, bodyLengthBetween } from '../common/assertions.js'
+import {
+	isEnvValid,
+	isTestEnabledOnEnv,
+	DEV,
+	UAT,
+	PROD,
+} from '../common/envs.js'
+import dotenv from 'k6/x/dotenv'
+
+const REGISTERED_ENVS = [DEV, UAT, PROD]
+
+const services = JSON.parse(open('../../services/environments.json'))
+export let options = {}
+let params = {}
+let baseUrl
+let myEnv
+
+if (isEnvValid(__ENV.TARGET_ENV)) {
+	myEnv = dotenv.parse(open(`../../.env.${__ENV.TARGET_ENV}.local`))
+	baseUrl = services[`${__ENV.TARGET_ENV}_issuer`].baseUrl
+
+	options.tlsAuth = [
+		{
+			domains: [baseUrl],
+			cert: open(`../../certs/${myEnv.MAUTH_CERT_NAME}`),
+			key: open(`../../certs/${myEnv.MAUTH_PRIVATE_KEY_NAME}`),
+		},
+	]
+
+	params.headers = {
+		'Ocp-Apim-Subscription-Key': myEnv.APIM_RTDPRODUCT_SK,
+		'Ocp-Apim-Trace': 'true',
+	}
+}
+
+export default () => {
+	if (
+		!isEnvValid(__ENV.TARGET_ENV) ||
+		!isTestEnabledOnEnv(__ENV.TARGET_ENV, REGISTERED_ENVS)
+	) {
+		return
+	}
+	group('Payment Instrument API', () => {
+		group('Should get hashed pans', () =>
+			assert(getHashedPan(baseUrl, params), [
+				statusOk(),
+				bodyLengthBetween(0, myEnv.RTD_HASHPAN_MAX_CONTENT_LENGTH),
+			])
+		)
+		group('Should get salt', () =>
+			assert(getSalt(baseUrl, params), [statusOk()])
+		)
+	})
+}
